custom_assets/requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
pymongo==3.12.0
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.2.4
